requirements.txt
pandas
assemblyai
pympi-ling
openpyxl
werpy
pyannote.metrics